Merge in the upgrade to vc140, plus number of new features and warning fixes.

currently at a stable state.

See merge request !25

+++ Pulp/Common/Logging/LoggerBase.cpp
@@ -23,7 +23,7 @@ void LoggerBase::SetParent(ILog* pLog)
 /// ----------------------------------------------------------------------------------
 
 void LoggerBase::Log(const SourceInfo& sourceInfo, const char* channel,
-	size_t verbosity, const char* format, va_list args)
+	int verbosity, const char* format, va_list args)
 {
 	DoLog(sourceInfo, channel, verbosity, format, args);
 }
@@ -63,7 +63,7 @@ void LoggerBase::AssertVariable(const SourceInfo& sourceInfo,
 
 
 void LoggerBase::DoLog(const SourceInfo& sourceInfo, const char* channel,
-	size_t verbosity, const char* format, va_list args)
+	int verbosity, const char* format, va_list args)
 {
 	X_UNUSED(sourceInfo);
 	X_UNUSED(channel);
